Version update: 7.4.3 -> 7.6.2
Also add a script for easier generation of the list of packages
available via Debian.

Also remove pip-selfcheck.json. Did not find any clear description of
what this is but it seems to be used to warn the user when using an
old version of pip. Committing it to the repo did not seem useful.

Bug: T105374
Change-Id: Iac82ea202856315ba187c852070a4337bd7781a6